Move to new apb repository organization and update riscv

+++ ips/scripts/vcompile_apb_gpio.sh
@@ -1,8 +1,19 @@
 #!/bin/tcsh
-
 source scripts/colors.sh
 
 echo "${Green}--> Compiling APB GPIO INTERFACE... ${NC}"
+
+# check if environment variables are defined
+if (! $?MSIM_LIBS_PATH ) then
+  echo "${Red} MSIM_LIBS_PATH is not defined ${NC}"
+  exit 1
+endif
+
+if (! $?IPS_PATH ) then
+  echo "${Red} IPS_PATH is not defined ${NC}"
+  exit 1
+endif
+
 
 echo "${Green}library: apb_gpio_lib ${NC}"
 rm -rf ${MSIM_LIBS_PATH}/apb_gpio_lib
@@ -13,7 +24,7 @@ vmap apb_gpio_lib ${MSIM_LIBS_PATH}/apb_gpio_lib
 echo "${Green}Compiling component:   ${Brown} axi_gpio ${NC}"
 echo "${Red}"
 
-vlog -work apb_gpio_lib -quiet -sv ${IPS_PATH}/apb/apb_gpio/apb_gpio.sv    || exit 1
+vlog -work apb_gpio_lib -quiet -sv ${IPS_PATH}/apb_peripherals/apb_gpio/apb_gpio.sv    || exit 1
 
 echo "${Cyan}--> APB GPIO INTERFACE compilation complete! ${NC}"
